+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ### Changes
 
-- removed page from the folder names inside of pages folder
-- changed fightersTable from using material ui to using bootstrap table
+- added row event to fighters table to go to specific fighter page
+- remvod pagination component, using bootstrap table pagination
 
 ### Additions
 
--
+- added fighter page